requirements.txt
amqp==5.2.0
attrs==23.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
blinker==1.8.2
celery==5.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
exceptiongroup==1.2.1
fake-useragent==1.5.1
Flask==3.0.3
h11==0.14.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kombu==5.3.7
lxml==5.2.2
MarkupSafe==2.1.5
nltk==3.8.1
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
prompt_toolkit==3.0.47
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
regex==2024.5.15
requests==2.32.3
Sastrawi==1.0.1
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.1
tzdata==2024.1
urllib3==1.26.18
uvicorn==0.30.1
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.1
Werkzeug==3.0.3
wsproto==1.2.0
XlsxWriter==3.2.0
zipp==3.19.2
